bunch of small random code updates lowering code size another 500-600 bytes

+++ i_quit.c
@@ -73,7 +73,7 @@ void RestoreRealTimeClock(void);
 
 extern boolean mousepresent;
 extern boolean usemouse;
-int32_t I_ResetMouse(void);
+int16_t I_ResetMouse(void);
 
 
 /*---------------------------------------------------------------------
